feat: server-side heartbeat cache, JSON tree viewer, Mongo/UI polish
- HeartbeatMonitor: singleton background service caching heartbeat per
  env; polling pauses when no subscribers, cache released after
  IdleReleaseMinutes (configurable); fires OnServicesUpdated event
- ServiceMetricsMonitor: uses IHeartbeatMonitor instead of calling
  IHeartbeatService directly; metrics cache (file offsets + history)
  survives idle window for fast resume without re-reading files
- ServicesPage: removed 15-second client-side PeriodicTimer; subscribes
  to HeartbeatMonitor.OnServicesUpdated for table refresh instead
- HomeMemoryTreemap: LoadAsync is now synchronous (reads from cache);
  subscribes to both OnServicesUpdated and OnMetricsUpdated
- appsettings: IdleReleaseMinutes added to Heartbeats and Logs sections
- MongoService.GetCollectionsAsync: parallel $collStats aggregation
  instead of sequential collStats commands — much faster on large DBs
- MongoDashboard: removed AvgDocSize column (still in info dialog);
  Indexes column now sortable; filter extended with docs/storage/indexes
- MongoCollectionInspector: removed Timestamp column; _id label in
  detail panel header; filter hint updated to mention _id search
- JSON viewer: new js/json-viewer.js renders collapsible, syntax-coloured
  JSON trees; replaces <pre> in Mongo and Kafka detail panels;
  dark/light theme token set in CSS
- NLog layout: callsite moved after message, shows ClassName.Method
- Services card width: 220px → 240px
- Rows-per-page selector: added 100/200 options; width 68→90px

+++ BatchMonitor_Step1/src/NxtUI.Core/Models/MongoModels.cs
@@ -2,37 +2,37 @@ namespace NxtUI.Core.Models;
 
 public record MongoDatabaseInfo
 {
-    public string Name            { get; init; } = string.Empty;
-    public int    CollectionCount { get; init; }
-    public long   SizeBytes       { get; init; }
+    public string Name { get; init; } = string.Empty;
+    public long CollectionCount { get; init; }
+    public long SizeBytes { get; init; }
 }
 
 public record MongoCollectionSummary
 {
-    public string Name             { get; init; } = string.Empty;
-    public long   DocumentCount    { get; init; }
-    public long   AvgDocSizeBytes  { get; init; }
-    public long   StorageSizeBytes { get; init; }
-    public int    IndexCount       { get; init; }
+    public string Name { get; init; } = string.Empty;
+    public long DocumentCount { get; init; }
+    public long AvgDocSizeBytes { get; init; }
+    public long StorageSizeBytes { get; init; }
+    public int IndexCount { get; init; }
 }
 
 public record MongoDocument
 {
-    public string    Id        { get; init; } = string.Empty;
-    public string    Json      { get; init; } = string.Empty;
+    public string Id { get; init; } = string.Empty;
+    public string Json { get; init; } = string.Empty;
     public DateTime? Timestamp { get; init; }
 }
 
 public record MongoIndexInfo
 {
-    public string Name   { get; init; } = string.Empty;
-    public string Keys   { get; init; } = string.Empty;
-    public bool   Unique { get; init; }
-    public bool   Sparse { get; init; }
+    public string Name { get; init; } = string.Empty;
+    public string Keys { get; init; } = string.Empty;
+    public bool Unique { get; init; }
+    public bool Sparse { get; init; }
 }
 
 public record MongoCollectionDetails
 {
-    public MongoCollectionSummary        Summary { get; init; } = new();
+    public MongoCollectionSummary Summary { get; init; } = new();
     public IReadOnlyList<MongoIndexInfo> Indexes { get; init; } = [];
 }
